Add unit tests, fix schemaless resource properties
The 'node' property of the Chef Solo resource is a schemaless map.
Previously, this was handled by simply assuming that whatever was
inside was plain data. Heat actually allows intrinsic functions to be
within these maps. This commit fixes that by using an extension of the
HOT.Primitive type that wraps the elements of lists and maps with
ResourcePropertyWrappers.